Add the possibility to override the BLUETOOTH option for the X9D+2019.
Some Taranis X9D+2019 transmitters were sold without a bluetooth
module. Preparing a cmake Makefile for building the firmware is now
possible with "-DBLUETOOTH=NO". It's not a huge issue, but it's
low-hanging fruit for saving 5.4 kByte on a X9D+2019.

This change keeps the standard build for an X9D+2019 backwards
compatible, as bluetooth remains compiled in by default.

Example with -DBLUETOOTH=YES (or default, not specified at all):

  Flash:   381.27   74.47%     512       130.73
  CCRAM:    39.44   61.63%      64        24.56
  SRAM1:    29.71   26.52%     112        82.29
  SRAM2:     0.00    0.00%      16        16.00

Example without -DBLUETOOTH=NO:

  Flash:   375.87   73.41%     512       136.13
  CCRAM:    39.01   60.96%      64        24.99
  SRAM1:    29.71   26.52%     112        82.29
  SRAM2:     0.00    0.00%      16        16.00